Individual Optimiser passes
Now remove post return with no way of getting there ops too
