+++ ulox/ulox.core/Package/Runtime/Optimiser/IOptimiserPass.cs
@@ -1,0 +1,8 @@
+﻿namespace ULox
+{
+    public interface IOptimiserPass
+    {
+        void Reset();
+        void Run(Optimiser optimiser, CompiledScript compiledScript);
+    }
+}
